Docker: Increase isolation by running as a non-root user
Using a dedicated user helps avoid container-escape issues such as
CVE-2019-5736.

Much of the complexity arises from continuing to use port 80/443 for web
serving.  As the port mapping is arbitrary and YAGPDB already supports running
on a non-privileged port, it is trivial to eliminate that complexity. However,
existing users may have their own configs that make assumptions about the
exposed port.

The app-as-entrypoint is somewhat convenient for decreasing verbosity of CMD
configuration, but prevented the necessary config to implement this and is
against best practices:
https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#cmd

As YAGPDB now references this image, this should be submitted as close to the
corresponding pull request. TODO: Add link